Implemented pedantic formatting rules

+++ seeds/lib/services/library/manager.dart
@@ -1,13 +1,16 @@
 import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:seeds/services/library/library.dart';
-import 'package:xml/xml.dart';
 import 'package:http/http.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:xml/xml.dart';
+
+import 'library.dart';
 
 class LibraryManager extends ChangeNotifier {
-  static final String _webStorageURL = 'https://firebasestorage.googleapis.com/v0/b/scripture-seeds.appspot.com/o/';
+  static final String _webStorageURL =
+      'https://firebasestorage.googleapis.com/v0/b/scripture-seeds.appspot.com/o/';
 
   final String lang;
   final AssetBundle assets;
@@ -18,11 +21,8 @@ class LibraryManager extends ChangeNotifier {
   DateTime _lastRefresh;
   DateTime get lastRefresh => _lastRefresh;
 
-  LibraryManager({
-    this.assets,
-    this.daysBetweenRefresh = 2,
-    this.lang = 'eng'
-  }) : _libFileName = 'library_$lang' {
+  LibraryManager({this.assets, this.daysBetweenRefresh = 2, this.lang = 'eng'})
+      : _libFileName = 'library_$lang' {
     _loadLibrary().then((library) async {
       // Update library and notify listeners
       if (library != null) {
@@ -30,60 +30,65 @@ class LibraryManager extends ChangeNotifier {
         notifyListeners();
       }
 
-      // If library was unable to load, or if cache is over 48 hours old, download from web
-      if (library == null || await _shouldRefreshCache())
-        refreshLibrary();
+      // If library was unable to load, or if cache is over 48 hours old,
+      // download from web.
+      if (library == null || await _shouldRefreshCache()) {
+        await refreshLibrary();
+      }
     });
   }
 
   // Attempt to refresh library from web.
   Future<void> refreshLibrary() async {
     print('Refreshing library...');
-    Library newLibrary = Library.fromXmlElement((await _loadFromWeb())?.rootElement);
+    var newLibrary =
+        Library.fromXmlElement((await _loadFromWeb())?.rootElement);
 
     if (newLibrary == null) return;
 
     if (newLibrary.version > (library?.version ?? -1)) {
       library = newLibrary;
-    } else
+    } else {
       print('Web library is outdated.');
+    }
 
     _lastRefresh = DateTime.now();
     notifyListeners();
 
-    _updateCacheFromDownload();
+    await _updateCacheFromDownload();
   }
 
   /// Cache Management Functions
   // Gets cache file
   Future<File> _getCacheFile({bool newDownload = false}) async {
-    Directory tempDirectory = await getTemporaryDirectory();
-    return File(tempDirectory.path + '/lib_cache/$_libFileName' +
-                (newDownload ? '_new' : '') + '.xml');
+    var tempDirectory = await getTemporaryDirectory();
+    return File(
+        '${tempDirectory.path}/lib_cache/$_libFileName${newDownload ? '_new' : ''}.xml');
   }
 
   // Swaps new cache file for old cache file
   Future<void> _updateCacheFromDownload() async {
-    File newCache = await _getCacheFile(newDownload: true);
-    File oldCache = await _getCacheFile();
+    var newCache = await _getCacheFile(newDownload: true);
+    var oldCache = await _getCacheFile();
 
     if (await newCache.exists()) {
       print('Saving library to cache...');
 
-      if (await oldCache.exists())
-        await oldCache.delete();
+      if (await oldCache.exists()) await oldCache.delete();
 
       await newCache.rename(oldCache.path);
     }
   }
 
-  // Returns true if no cache is present or if cache is over daysBetweenRefresh days old
+  // Returns true if no cache is present or if cache is over
+  // daysBetweenRefresh days old
   Future<bool> _shouldRefreshCache() async {
-    File cache = await _getCacheFile();
+    var cache = await _getCacheFile();
 
     if (await cache.exists()) {
-      DateTime lastRefresh = await cache.lastModified();
-      return DateTime.now().difference(lastRefresh).inDays >= daysBetweenRefresh;
+      var lastRefresh = await cache.lastModified();
+      return DateTime.now().difference(lastRefresh).inDays >=
+          daysBetweenRefresh;
     }
 
     return true;
@@ -95,17 +100,15 @@ class LibraryManager extends ChangeNotifier {
     print('Loading local library...');
 
     // Load from cache and assets
-    List<Library> libs = await Future.wait([
-      _loadFromCache().then(
-        (libDoc) => Library.fromXmlElement(libDoc?.rootElement)
-      ),
-      _loadFromAssets().then(
-        (libDoc) => Library.fromXmlElement(libDoc?.rootElement)
-      )
+    var libs = await Future.wait([
+      _loadFromCache()
+          .then((libDoc) => Library.fromXmlElement(libDoc?.rootElement)),
+      _loadFromAssets()
+          .then((libDoc) => Library.fromXmlElement(libDoc?.rootElement))
     ]);
 
-    Library cacheLibrary = libs[0];
-    Library assetsLibrary = libs[1];
+    var cacheLibrary = libs[0];
+    var assetsLibrary = libs[1];
 
     // If cache is older than asset, return asset
     if ((assetsLibrary?.version ?? -1) > (cacheLibrary?.version ?? -1)) {
@@ -130,22 +133,24 @@ class LibraryManager extends ChangeNotifier {
     if (assets != null) {
       print('Loading library from assets...');
       try {
-        return XmlDocument.parse(await assets.loadString('assets/$_libFileName.xml'));
+        return XmlDocument.parse(
+            await assets.loadString('assets/$_libFileName.xml'));
+        // ignore: avoid_catching_errors
       } on FlutterError {
         print('Could not find library asset for language "$lang"');
       } on XmlParserException catch (e) {
         print('Parsing of library XML from assets failed: $e');
       }
-    }
-    else
+    } else {
       print('No assets provided to library file manager.');
+    }
 
     return null;
   }
 
   // Loads library file from cache
   Future<XmlDocument> _loadFromCache() async {
-    File cache = await _getCacheFile();
+    var cache = await _getCacheFile();
 
     if (await cache.exists()) {
       _lastRefresh = await cache.lastModified();
@@ -166,7 +171,7 @@ class LibraryManager extends ChangeNotifier {
 
   // Loads library file from web and saves to cache
   Future<XmlDocument> _loadFromWeb() async {
-    File libFile = await _downloadLibrary();
+    var libFile = await _downloadLibrary();
 
     if (libFile != null) {
       print('Loading library from web download...');
@@ -190,14 +195,13 @@ class LibraryManager extends ChangeNotifier {
   /// Download Function
   // Downloads the library and saves to the cache
   Future<File> _downloadLibrary() async {
-    Uri url = Uri.parse('$_webStorageURL$_libFileName.xml?alt=media');
+    var url = Uri.parse('$_webStorageURL$_libFileName.xml?alt=media');
 
     print('Downloading library from "$url"...');
-    Response response = await get(url);
+    var response = await get(url);
     if (response.statusCode == 200) {
-      File cache = await _getCacheFile(newDownload: true);
-      if (!await cache.exists())
-        await cache.create(recursive: true);
+      var cache = await _getCacheFile(newDownload: true);
+      if (!await cache.exists()) await cache.create(recursive: true);
       await cache.writeAsBytes(response.bodyBytes);
 
       return cache;
